pyproject.toml
[tool.poetry]
name = "toll-roads-backend"
version = "0.1.0"
description = "API for toll-roads project"
authors = ["Mihail <infast1k@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-rest-framework = "^0.1.0"
django-cors-headers = "^4.4.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
redis = "^5.0.8"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
